ORC-1130:[C++] add a reset function in the ByteRleEncoderImpl and add the test coverage for seek
