+++ c++/src/ByteRLE.cc
@@ -63,6 +63,11 @@ namespace orc {
 
     virtual void suppress() override;
 
+    /**
+     * Reset to initial state
+     */
+    void reset();
+
   protected:
     std::unique_ptr<BufferedOutputStream> outputStream;
     char* literals;
@@ -82,12 +87,7 @@ namespace orc {
                                 std::unique_ptr<BufferedOutputStream> output)
                                   : outputStream(std::move(output)) {
     literals = new char[MAX_LITERAL_SIZE];
-    numLiterals = 0;
-    tailRunLength = 0;
-    repeat = false;
-    bufferPosition = 0;
-    bufferLength = 0;
-    buffer = nullptr;
+    reset();
   }
 
   ByteRleEncoderImpl::~ByteRleEncoderImpl() {
@@ -205,15 +205,19 @@ namespace orc {
     recorder->add(static_cast<uint64_t>(numLiterals));
   }
 
-  void ByteRleEncoderImpl::suppress() {
-    // written data can be just ignored because they are only flushed in memory
-    outputStream->suppress();
-
+  void ByteRleEncoderImpl::reset() {
     numLiterals = 0;
     tailRunLength = 0;
     repeat = false;
     bufferPosition = 0;
     bufferLength = 0;
+    buffer = nullptr;
+  }
+
+  void ByteRleEncoderImpl::suppress() {
+    // written data can be just ignored because they are only flushed in memory
+    outputStream->suppress();
+    reset();
   }
 
   std::unique_ptr<ByteRleEncoder> createByteRleEncoder
